#BDRE-200 added adq table in all db's

+++ md-dao/src/main/java/com/wipro/ats/bdre/md/dao/AppDeploymentQueueDAO.java
@@ -94,7 +94,7 @@ public class AppDeploymentQueueDAO {
         }
     }
 
-    public void delete(Integer id) {
+    public void delete(Long id) {
         Session session = sessionFactory.openSession();
         try {
             session.beginTransaction();
